Add decoding and printing addresses work.
This now supports cashaddress, legacy addresses and it prints
basic info about the usages of said address.

+++ indexer/AddressIndexer.h
@@ -42,7 +42,7 @@ public:
 
     struct TxData {
         int offsetInBlock = 0;
-        short blockHeight = -1;
+        int blockHeight = -1;
         short outputIndex = -1;
     };
